+++ NEWS.md
@@ -13,6 +13,8 @@ annotating/highlighting pairwise comparisons between levels of a factor mapped
 to the _x_ aesthetic. They can also be used to highlight ranges of values
 when the variable mapped to _x_ is a continuous numerical vector. The user
 interfaces are consistent with those of `geom_text_s()` and `geom_label_s()`.
+-   Improve handling of nudge values when dates or times are mapped to _x_ or
+_y_.
 
 # ggpp 0.5.3